+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Added Tile attribute: allStagedCal - Returns all staged calibration coefficients.
   * Added Tile attribute: allLiveCal - Returns all live calibration coefficients.
   * Added Tile command: LoadCalibrationCoefficientsForChannels - Loads calibration coefficients for specified channels.
+* [THORN-372] Added command  'LoadCalibrationCoefficientForChannels ' to MccsTile & SpsStation
 * [THORN-389] Bumping TPM FPGA firmware version to 9.0.0
   * Increased bitwidth of subarray and substation ID
   * Calibration coefficients are automatically copied back when banks are switched